Add config .env file and logger and exception

+++ src/events/events.controller.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line prettier/prettier
-import { Controller, Get, Post, Patch, Delete, Param, Body, HttpCode } from '@nestjs/common';
+import { Controller, Get, Post, Patch, Delete, Param, Body, HttpCode, Logger, NotFoundException } from '@nestjs/common';
 import { CreateEventDto } from './create-event.dto';
 import { UpdateEventDto } from './update-event.dto';
 import { Event } from './event.entity';
@@ -9,13 +9,18 @@ import { InjectRepository } from '@nestjs/typeorm';
 @Controller('/events')
 export class EventsController {
   // private events: Event[] = [];
+
+  private readonly logger = new Logger(EventsController.name);
   constructor(
     @InjectRepository(Event)
     private readonly repository: Repository<Event>,
   ) {}
   @Get()
   async findAll() {
-    return this.repository.find();
+    this.logger.log(`Hit the findAll route`);
+    const events = await this.repository.find();
+    this.logger.debug(`Found ${events.length} events`);
+    return events;
   }
 
   // order of methods in class mater!
@@ -37,8 +42,12 @@ export class EventsController {
 
   @Get(':id')
   async findOne(@Param('id') id) {
-    console.log(typeof id);
-    return await this.repository.findOne({ where: { id: id } });
+    // console.log(typeof id);
+    const event = await this.repository.findOne({ where: { id: id } });
+    if (!event) {
+      throw new NotFoundException();
+    }
+    return event;
   }
 
   @Post()
@@ -51,7 +60,10 @@ export class EventsController {
 
   @Patch(':id')
   async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOne({ where: { id: id } });
+    if (!event) {
+      throw new NotFoundException();
+    }
 
     return await this.repository.save({
       ...event,
@@ -63,7 +75,10 @@ export class EventsController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id') id) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOne({ where: { id: id } });
+    if (!event) {
+      throw new NotFoundException();
+    }
     await this.repository.remove(event);
   }
 }
